新增崩溃保存，以及信息打印

+++ SamSoarII.Utility/TempDebugger.cs
@@ -8,14 +8,13 @@ namespace SamSoarII.Utility
 {
     public class TempDebugger
     {
-        public static string DebugPath = FileHelper.AppRootPath + @"\Debug.txt";
+        public static string DebugPath = FileHelper.AppRootPath + @"\Temp\Debug.txt";
         private static StreamWriter writer;
         static TempDebugger()
         {
-            if (!File.Exists(DebugPath))
-            {
-                File.Create(DebugPath);
-            }
+            if (!Directory.Exists(FileHelper.AppRootPath + @"\Temp"))
+                Directory.CreateDirectory(FileHelper.AppRootPath + @"\Temp");
+            if (!File.Exists(DebugPath)) File.Create(DebugPath);
             writer = new StreamWriter(DebugPath);
         }
         public static void WriteLine(string message)
@@ -30,12 +29,18 @@ namespace SamSoarII.Utility
             writer.Flush();
             //writer.Close();
         }
+        public static void WriteLine()
+        {
+            writer.WriteLine();
+            writer.Flush();
+        }
         public static void Close()
         {
             writer.Close();
         }
         public static void Dispose()
         {
+            writer.Close();
             writer.Dispose();
         }
     }
